feat: Handle Korean input with keypress event
Due to the occurrence of two keyup events when pressing Enter after inputting Korean, and to ensure accurate handling of input, this commit introduces the use of the keypress event instead of keyup. This change is made to address the specific issue where two keyup events are triggered and to handle the scenario more appropriately. Additionally, a check for isComposing===false has been added to ensure proper handling of the keypress event.